finalização commit

+++ src/main/java/entities/HourContract.java
@@ -10,7 +10,7 @@ import java.util.Date;
 public class HourContract {
     private Date date;
     private Double valuePerHour;
-    private Integer hours;
+    private Double hours;
 
     public HourContract() {
     }
